custom app composition with H5Manager
