+++ packages/pymodaq_gui/src/pymodaq_gui/utils/custom_app.py
@@ -1,3 +1,4 @@
+import inspect
 from pathlib import Path
 from typing import Union, TYPE_CHECKING, Dict, Optional
 
@@ -6,31 +7,20 @@ from qt_themes import Theme
 from qtpy.QtCore import QObject, QLocale
 from qtpy import QtCore, QtWidgets
 
+from pymodaq_gui.managers.h5manager import FileStatus, H5Manager
 from pymodaq_utils.config import GlobalConfig as Config
 from pymodaq_utils.logger import set_logger, get_module_name
 from pymodaq_utils.config import get_set_path, get_set_local_dir
 from pymodaq_utils.warnings import deprecation_msg
-from pymodaq_utils.enums import BaseEnum
-from pymodaq_gui.utils.widgets import QLED
+
 from pymodaq_gui.utils.dock import DockArea, Dock
 from pymodaq_gui.managers.action_manager import ActionManager
 from pymodaq_gui.managers.parameter_manager import ParameterManager
 from pymodaq_gui.parameter import ParameterTree
 from pymodaq_gui.utils.splash import get_splash_sc
-from pymodaq_gui.utils import select_file
-
-from pymodaq_gui.h5modules.saving import H5Saver
 
 logger = set_logger(get_module_name(__file__))
 config = Config()
-
-
-class FileStatus(BaseEnum):
-    NEW = 0
-    REOPENED = 1
-    CLOSED = 2
-    REOPENED_ANOTHER = 3
-    NO_FILE = 4
 
 
 class CustomApp(QObject, ActionManager, ParameterManager):
@@ -117,8 +107,8 @@ class CustomApp(QObject, ActionManager, ParameterManager):
     """
 
     log_signal = QtCore.Signal(str)
-    _show_h5file_statusbar_widgets = False
-    _h5_base_group_name = 'AppData'  # rename that in your app/extension to give a meaningful name to your base group
+    show_h5file_statusbar_widgets = False
+    h5_base_group_name = 'AppData'  # rename that in your app/extension to give a meaningful name to your base group
     params = []
 
     def __init__(self, parent: Union[DockArea, QtWidgets.QMainWindow, QtWidgets.QWidget] = None,
@@ -151,7 +141,7 @@ class CustomApp(QObject, ActionManager, ParameterManager):
 
         self._title: str = ''
         self.title = title
-        self._h5saver: H5Saver = None  # to use only if you want to save data,
+
         # then call self.h5saver property
         self.docks: Dict[str, Dock] = dict([])
 
@@ -179,6 +169,35 @@ class CustomApp(QObject, ActionManager, ParameterManager):
                           self.__class__.__name__,
                           self.menubar if self.mainwindow is not None else None)
 
+        self._h5_manager = H5Manager(self)
+
+        self._delegates = [self._h5_manager,]  # to deal with eventual other composed objects
+
+    def __getattr__(self, name):
+        """ To access composed objects attributes and methods"""
+        delegates = self.__dict__.get("_delegates", [])
+        for delegate in delegates:
+            if name in dir(delegate) or hasattr(type(delegate), name):  # to prevent executing some code and check for errors
+                try:
+                    return getattr(delegate, name)
+                except AttributeError as e:
+                    raise e
+
+        raise AttributeError(f"'{type(self).__name__}' object has no attribute '{name}'")
+
+    def __dir__(self):
+        """ to expose the composed objects attributes and methods to autocompletion tools"""
+        attributes = set(super().__dir__())
+
+        # Add public attributes
+        delegates = self.__dict__.get("_delegates", [])
+        for delegate in delegates:
+            for attr in dir(delegate):
+                if not attr.startswith('_'):  # skip private and dunder methods
+                    attributes.add(attr)
+        return sorted(list(attributes))
+
+
     @classmethod
     def get_local_folder(cls, user=False) -> Path:
         """ Create a local User or system wide folder to store things about this extension"""
@@ -202,7 +221,7 @@ class CustomApp(QObject, ActionManager, ParameterManager):
 
         self.insert_custom_status_widgets()
 
-        if self._show_h5file_statusbar_widgets:
+        if self.show_h5file_statusbar_widgets:
             self.insert_h5stuff_status()
 
     def set_permanent_status(self, status: str):
@@ -355,180 +374,3 @@ class CustomApp(QObject, ActionManager, ParameterManager):
         To be reimplemented
         """
         pass
-
-    def insert_h5stuff_status(self):
-        self._file_open_LED = QLED()
-        self._file_open_LED.set_as_false()
-        self._file_open_LED.clickable = False
-        self._file_open_LED.setToolTip('H5 file open and accessible')
-
-        self._swmr_label = QtWidgets.QLabel('')
-        self._swmr_label.setToolTip('SWMR mode status')
-        self._swmr_label.setVisible(False)
-
-        self.statusbar.addPermanentWidget(QtWidgets.QLabel('File:'))
-        self.statusbar.addPermanentWidget(self._file_open_LED)
-        self.statusbar.addPermanentWidget(self._swmr_label)
-
-    @property
-    def h5saver(self) -> H5Saver:
-        if self._h5saver is None:
-            self._h5saver = H5Saver()
-            self._h5saver.settings.child('do_save').hide()
-            self._h5saver.settings.child('custom_name').hide()
-            self._h5saver.settings['base_name'] = self._h5_base_group_name
-            self._h5saver.new_file_sig.connect(self.create_new_file)
-            self._h5saver.file_changed_sig.connect(self.update_file_status_led)
-
-        status = self.open_file()
-        if status == FileStatus.NO_FILE:
-            self.create_new_file(True)
-        return self._h5saver
-
-    @QtCore.Slot(bool)
-    def create_new_file(self, new_file):
-        """ Slot of the New File button in the H5Saver settings Tree"""
-
-        if new_file:
-            self.close_file()
-            # Explicitly create a new file (don't reopen existing)
-            try:
-                self._h5saver.init_file(update_h5=True)
-                logger.info(f"Created new h5 file: {self._h5saver.settings['current_h5_file']}")
-            except Exception as e:
-                logger.error(f"Could not create new h5 file: {e}")
-
-    def open_file(self) -> FileStatus:
-        """ Try to reopen the current h5 file if it is closed.
-        """
-        if self._h5saver is not None and not self._h5saver.isopen():
-            current_file = self._h5saver.settings['current_h5_file']
-            if current_file and Path(current_file).exists():
-                return self._try_open_existing_file(current_file)
-            else:
-                return FileStatus.NO_FILE
-        return FileStatus.REOPENED
-
-    def close_file(self):
-        self._h5saver.close_file()
-
-    def _try_open_existing_file(self, current_file: str | Path) -> FileStatus:
-        """Try to open an existing file, asking user what to do if locked.
-
-        Return:
-        -------
-        FileStatus
-        """
-        while True:
-            try:
-                logger.debug(f"Reopening existing h5 file: {current_file}")
-                self._h5saver.init_file(addhoc_file_path=current_file)
-                return FileStatus.REOPENED  # Success
-            except Exception as e:
-                if 'lock' in str(e).lower() or 'errno = 0' in str(e).lower():
-                    # File is locked - ask user what to do
-                    msg = QtWidgets.QMessageBox()
-                    msg.setIcon(QtWidgets.QMessageBox.Icon.Warning)
-                    msg.setWindowTitle("File Locked")
-                    msg.setText(f"Cannot open file:\n{current_file}\n\n"
-                                f"The file may be open in another application.")
-                    msg.setInformativeText("Close the file elsewhere and click Retry, "
-                                           "or select a different file.")
-                    retry_btn = msg.addButton("Retry", QtWidgets.QMessageBox.ButtonRole.ActionRole)
-                    new_auto_btn = msg.addButton("New File (Auto)", QtWidgets.QMessageBox.ButtonRole.AcceptRole)
-                    browse_btn = msg.addButton("Browse...", QtWidgets.QMessageBox.ButtonRole.ActionRole)
-                    msg.addButton(QtWidgets.QMessageBox.StandardButton.Cancel)
-                    msg.exec()
-
-                    if msg.clickedButton() == retry_btn:
-                        continue  # Try again
-                    elif msg.clickedButton() == new_auto_btn:
-                        logger.info("User chose to create new file (auto)")
-                        self._h5saver.init_file(update_h5=True)
-                        return FileStatus.NEW
-                    elif msg.clickedButton() == browse_btn:
-                        # Let user select an existing file to append to
-                        file_path, _ = QtWidgets.QFileDialog.getOpenFileName(
-                            None, "Select HDF5 File",
-                            str(Path(current_file).parent),
-                            "HDF5 Files (*.h5);;All Files (*)",
-                        )
-                        if file_path:
-                            logger.info(f"User selected file: {file_path}")
-                            try:
-                                self._h5saver.init_file(addhoc_file_path=file_path)
-                                return FileStatus.REOPENED_ANOTHER
-                            except Exception as e2:
-                                logger.warning(f"Could not open selected file: {e2}")
-                                continue  # Show dialog again
-                        else:
-                            continue  # User cancelled browse, show dialog again
-                    else:
-                        # User cancelled - leave h5_file unchanged
-                        logger.info("User cancelled file selection - keeping current file state")
-                        return FileStatus.CLOSED
-                else:
-                    # Other error - fall back to new file
-                    logger.warning(f"Could not reopen h5 file: {e}")
-                    self._h5saver.init_file(update_h5=True)
-                    return FileStatus.NEW
-
-    def load_file(self):
-        self.h5saver.load_file(self.h5saver.h5_file_path)
-        self.update_file_status_led()
-
-    def save_file(self):
-        Path(self.h5saver.settings['base_path']).mkdir(exist_ok=True)
-        filename = select_file(self.h5saver.settings['base_path'], save=True, ext='h5')
-        self.h5saver.h5_file.copy_file(str(filename), overwrite=True)
-
-    def set_file_open(self, is_open: bool):
-        """Update the file-open status LED.
-
-        Parameters
-        ----------
-        is_open:
-            True (green) if the h5 file is open and accessible, False (red) otherwise.
-        """
-        if self._show_h5file_statusbar_widgets:
-            self._file_open_LED.set_as(is_open)
-
-    def show_file_content(self):
-         if self._h5saver is not None:
-             self._h5saver.show_file_content()
-
-    def set_swmr_status(self, active: bool, compatible: bool = False):
-        """Show or hide the SWMR mode indicator in the status bar.
-
-        Parameters
-        ----------
-        active:
-            True if SWMR mode is currently active on the file.
-        compatible:
-            True if the file was created with SWMR support.
-        """
-        if self._show_h5file_statusbar_widgets:
-            if active:
-                self._swmr_label.setText('SWMR')
-                self._swmr_label.setToolTip('SWMR mode active')
-                self._swmr_label.setVisible(True)
-            elif compatible:
-                self._swmr_label.setText('SWMR file')
-                self._swmr_label.setToolTip('File created with SWMR support')
-                self._swmr_label.setVisible(True)
-            else:
-                self._swmr_label.setText('')
-                self._swmr_label.setToolTip('SWMR mode status')
-                self._swmr_label.setVisible(False)
-
-    def update_file_status_led(self):
-        """Reflect the current h5 file open/accessible state in the status bar LED
-        and the SWMR mode indicator."""
-
-        is_open = (self._h5saver is not None
-                   and self._h5saver.h5_file is not None
-                   and self._h5saver.isopen())
-        self.set_file_open(is_open)
-        swmr_active = is_open and self._h5saver.is_swmr_active
-        swmr_compatible = is_open and self._h5saver.is_swmr_compatible
-        self.set_swmr_status(swmr_active, swmr_compatible)
