+++ packages/pymodaq_gui/src/pymodaq_gui/utils/enums.py
@@ -1,0 +1,11 @@
+from pymodaq_utils.enums import StrEnum
+
+
+class MenuToolbarNames(StrEnum):
+    FILE = 'file'
+    SETTINGS = 'settings'
+    VIEW = 'view'
+    TOOLS = 'tools'
+    TOOLBARS = 'toolbars'
+    HELP = 'help'
+    RUNTIME = 'runtime'
